fix: Extracting global rulefile watcher ops into own file with stub impl for WASM
