+++ crates/ai/src/project_context/dummy_global_rules.rs
@@ -1,0 +1,22 @@
+use super::model::ProjectRule;
+use std::path::PathBuf;
+
+/// No-op stand-in for non-`local_fs` builds. File-based global rules require
+/// filesystem watchers that don't exist on WASM, so callers see an empty
+/// view here.
+#[derive(Debug, Default)]
+pub(crate) struct GlobalRules;
+
+impl GlobalRules {
+    pub(crate) fn active_rules(&self) -> impl Iterator<Item = ProjectRule> + '_ {
+        std::iter::empty()
+    }
+
+    pub(crate) fn paths(&self) -> impl Iterator<Item = PathBuf> + '_ {
+        std::iter::empty()
+    }
+
+    pub(crate) fn first_rule_parent(&self) -> Option<PathBuf> {
+        None
+    }
+}
