+++ crates/ai/src/project_context/mod.rs
@@ -1,1 +1,10 @@
+cfg_if::cfg_if! {
+    if #[cfg(feature = "local_fs")] {
+        mod global_rules;
+        pub(crate) use global_rules::GlobalRules;
+    } else {
+        mod dummy_global_rules;
+        pub(crate) use dummy_global_rules::GlobalRules;
+    }
+}
 pub mod model;
